index on feat/sg_GatherSettingView: ba99308 refactor: configure 인자변경..?
