+++ Letports/Letports/GoogleService-Info 2.plist
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>API_KEY</key>
+	<string>AIzaSyAVEW-T9Llpo0VjlrgNPKYzCkDUIiquyQM</string>
+	<key>GCM_SENDER_ID</key>
+	<string>350792606197</string>
+	<key>PLIST_VERSION</key>
+	<string>1</string>
+	<key>BUNDLE_ID</key>
+	<string>com.mosi.Letports</string>
+	<key>PROJECT_ID</key>
+	<string>letports-81f7f</string>
+	<key>STORAGE_BUCKET</key>
+	<string>letports-81f7f.appspot.com</string>
+	<key>IS_ADS_ENABLED</key>
+	<false></false>
+	<key>IS_ANALYTICS_ENABLED</key>
+	<false></false>
+	<key>IS_APPINVITE_ENABLED</key>
+	<true></true>
+	<key>IS_GCM_ENABLED</key>
+	<true></true>
+	<key>IS_SIGNIN_ENABLED</key>
+	<true></true>
+	<key>GOOGLE_APP_ID</key>
+	<string>1:350792606197:ios:94d6e4d965db49928a80bc</string>
+</dict>
+</plist>
